complete

+++ controllers/auth.js
@@ -119,7 +119,10 @@ exports.changeProfilePicture = (req, res, next) => {
     User.findById(req.user._id)
     .then(user => {
         if(user.profilePicture != "images/profile_default.png") {
-            fileHelper.deleteFile(user.profilePicture)
+            if(user.profilePicture)
+            {
+                fileHelper.deleteFile(user.profilePicture)
+            }
         }
         user.profilePicture = newProfilePicture;
         user.save();
